refactor jobs to a separate hook
